+++ libs/kah_gfx/src/gfx_converter.c
@@ -46,7 +46,7 @@ bool gfx_convert_shader_spv(const char *localAssetPath) {
 bool gfx_convert_texture_dds(char *localAssetPath) {
     const char *delim = (c_str_search_reverse(localAssetPath, "."));
     core_assert(c_str_search_reverse(localAssetPath, ".") != nullptr);
-    size_t copySize = delim - localAssetPath;
+    size_t copySize = (size_t)(delim - localAssetPath);
     char fileNameNoExt[256] = {};
     memcpy(fileNameNoExt, localAssetPath, copySize);
